+++ apps/nextjs-app/.size-limit.js
@@ -12,22 +12,22 @@ try {
 const pages = manifest.pages;
 
 const limitCfg = {
-  defaultSize: '120kb',
+  defaultSize: '80kb',
   pages: {
-    // Customize specific page limits if needed
-    '/_app': '160kb',
-    '/_error': '80kb',
-    '/404': '100kb',
     '/': '105kb',
-    '/demo': '105kb',
-    '/home': '100kb',
+    '/404': '70kb',
+    '/_app': '145kb',
+    '/_error': '80kb',
+    '/_monitor/sentry/csr-page': '75kb',
+    '/_monitor/sentry/ssr-page': '75kb',
+    '/home': '95kb',
   },
 };
 const getPageLimits = () => {
   let pageLimits = [];
   for (const [uri, paths] of Object.entries(pages)) {
     pageLimits.push({
-      name: `Browser page '${uri}'`,
+      name: `Page '${uri}'`,
       limit: limitCfg.pages?.[uri] ?? limitCfg.defaultSize,
       path: paths.map((p) => `.next/${p}`),
     });