+++ apps/remix-app/src/features/system/pages/__tests__/NotFoundPage.test.tsx
@@ -1,7 +1,7 @@
 import { NotFoundPage } from '@/features/system/pages';
 import { render, screen } from '@/test-utils';
 
-describe('NotFoundPage test', () => {
+describe('notFoundPage test', () => {
   it('should contains children', async () => {
     render(<NotFoundPage title={'404 - Not found'} />);
     expect(screen.getByTestId('not-found-title')).toHaveTextContent(