+++ packages/eslint-config-bases/src/bases/typescript.js
@@ -48,6 +48,7 @@ module.exports = {
     'linebreak-style': ['error', 'unix'],
     'no-empty-function': 'off',
     'import/default': 'off',
+    'import/no-duplicates': ['error', { considerQueryString: true }],
     'import/no-named-as-default-member': 'off',
     'import/no-named-as-default': 'off',
     'import/order': [
@@ -132,6 +133,8 @@ module.exports = {
           // DB / Graphql might use PascalCase for relationships
           'PascalCase',
         ],
+        leadingUnderscore: 'allowSingleOrDouble',
+        trailingUnderscore: 'allowSingleOrDouble',
       },
       {
         selector: ['typeAlias', 'interface'],